+++ types/byte.go
@@ -4,6 +4,8 @@ import (
 	"database/sql/driver"
 	"encoding/json"
 	"errors"
+
+	"github.com/volatiletech/sqlboiler/randomize"
 )
 
 // Byte is an alias for byte.
@@ -58,4 +60,13 @@ func (b *Byte) Scan(src interface{}) error {
 	}
 
 	return nil
+}
+
+// Randomize for sqlboiler
+func (b *Byte) Randomize(seed *randomize.Seed, fieldType string, shouldBeNull bool) {
+	if shouldBeNull {
+		*b = Byte(65) // Can't deal with a true 0-value
+	}
+
+	*b = Byte(seed.NextInt()%60 + 65) // Can't deal with non-ascii characters in some databases
 }