+++ types/pgeo/lseg.go
@@ -4,15 +4,19 @@ import (
 	"database/sql/driver"
 	"errors"
 	"fmt"
+
+	"github.com/volatiletech/sqlboiler/randomize"
 )
 
-//Line segments are represented by pairs of points that are the endpoints of the segment.
+// Lseg is a line segment and is represented by pairs of points that are the endpoints of the segment.
 type Lseg [2]Point
 
+// Value for the database
 func (l Lseg) Value() (driver.Value, error) {
 	return valueLseg(l)
 }
 
+// Scan from sql query
 func (l *Lseg) Scan(src interface{}) error {
 	return scanLseg(l, src)
 }
@@ -39,4 +43,13 @@ func scanLseg(l *Lseg, src interface{}) error {
 	*l = NewLseg(points[0], points[1])
 
 	return nil
+}
+
+func randLseg(seed *randomize.Seed) Lseg {
+	return Lseg([2]Point{randPoint(seed), randPoint(seed)})
+}
+
+// Randomize for sqlboiler
+func (l *Lseg) Randomize(seed *randomize.Seed, fieldType string, shouldBeNull bool) {
+	*l = randLseg(seed)
 }