+++ types/pgeo/nullCircle.go
@@ -2,13 +2,17 @@ package pgeo
 
 import (
 	"database/sql/driver"
+
+	"github.com/volatiletech/sqlboiler/randomize"
 )
 
+// NullCircle allows circle to be null
 type NullCircle struct {
 	Circle
 	Valid bool `json:"valid"`
 }
 
+// Value for database
 func (c NullCircle) Value() (driver.Value, error) {
 	if !c.Valid {
 		return nil, nil
@@ -17,6 +21,7 @@ func (c NullCircle) Value() (driver.Value, error) {
 	return valueCircle(c.Circle)
 }
 
+// Scan from sql query
 func (c *NullCircle) Scan(src interface{}) error {
 	if src == nil {
 		c.Circle, c.Valid = NewCircle(Point{}, 0), false
@@ -25,4 +30,15 @@ func (c *NullCircle) Scan(src interface{}) error {
 
 	c.Valid = true
 	return scanCircle(&c.Circle, src)
+}
+
+// Randomize for sqlboiler
+func (c *NullCircle) Randomize(seed *randomize.Seed, fieldType string, shouldBeNull bool) {
+	if shouldBeNull {
+		c.Valid = false
+		return
+	}
+
+	c.Valid = true
+	c.Circle = randCircle(seed)
 }