+++ types/pgeo/nullPoint.go
@@ -2,13 +2,17 @@ package pgeo
 
 import (
 	"database/sql/driver"
+
+	"github.com/volatiletech/sqlboiler/randomize"
 )
 
+// NullPoint allows point to be null
 type NullPoint struct {
 	Point
 	Valid bool `json:"valid"`
 }
 
+// Value for database
 func (p NullPoint) Value() (driver.Value, error) {
 	if !p.Valid {
 		return nil, nil
@@ -17,6 +21,7 @@ func (p NullPoint) Value() (driver.Value, error) {
 	return valuePoint(p.Point)
 }
 
+// Scan from sql query
 func (p *NullPoint) Scan(src interface{}) error {
 	if src == nil {
 		p.Point, p.Valid = NewPoint(0, 0), false
@@ -25,4 +30,15 @@ func (p *NullPoint) Scan(src interface{}) error {
 
 	p.Valid = true
 	return scanPoint(&p.Point, src)
+}
+
+// Randomize for sqlboiler
+func (p *NullPoint) Randomize(seed *randomize.Seed, fieldType string, shouldBeNull bool) {
+	if shouldBeNull {
+		p.Valid = false
+		return
+	}
+
+	p.Valid = true
+	p.Point = randPoint(seed)
 }