+++ types/pgeo/path.go
@@ -5,20 +5,24 @@ import (
 	"errors"
 	"fmt"
 	"regexp"
+
+	"github.com/volatiletech/sqlboiler/randomize"
 )
 
-//Paths are represented by lists of connected points.
-//Paths can be open, where the first and last points in the list are considered not connected,
+// Path is represented by lists of connected points.
+// Paths can be open, where the first and last points in the list are considered not connected,
 //or closed, where the first and last points are considered connected.
 type Path struct {
 	Points []Point
 	Closed bool `json:"closed"`
 }
 
+// Value for database
 func (p Path) Value() (driver.Value, error) {
 	return valuePath(p)
 }
 
+// Scan from sql query
 func (p *Path) Scan(src interface{}) error {
 	return scanPath(p, src)
 }
@@ -55,4 +59,13 @@ func scanPath(p *Path, src interface{}) error {
 	(*p).Closed = regexp.MustCompile(`^\(\(`).MatchString(val)
 
 	return nil
+}
+
+func randPath(seed *randomize.Seed) Path {
+	return Path{randPoints(seed, 3), newRandNum(seed) < 40}
+}
+
+// Randomize for sqlboiler
+func (p *Path) Randomize(seed *randomize.Seed, fieldType string, shouldBeNull bool) {
+	*p = randPath(seed)
 }