+++ types/pgeo/polygon.go
@@ -4,15 +4,19 @@ import (
 	"database/sql/driver"
 	"errors"
 	"fmt"
+
+	"github.com/volatiletech/sqlboiler/randomize"
 )
 
-//Polygons are represented by lists of points (the vertexes of the polygon).
+// Polygon is represented by lists of points (the vertexes of the polygon).
 type Polygon []Point
 
+// Value for database
 func (p Polygon) Value() (driver.Value, error) {
 	return valuePolygon(p)
 }
 
+// Scan from sql query
 func (p *Polygon) Scan(src interface{}) error {
 	return scanPolygon(p, src)
 }
@@ -37,4 +41,13 @@ func scanPolygon(p *Polygon, src interface{}) error {
 	}
 
 	return nil
+}
+
+func randPolygon(seed *randomize.Seed) Polygon {
+	return Polygon(randPoints(seed, 3))
+}
+
+// Randomize for sqlboiler
+func (p *Polygon) Randomize(seed *randomize.Seed, fieldType string, shouldBeNull bool) {
+	*p = randPolygon(seed)
 }